move utilities to util.py module
